requirements.txt
flask==1.1.2
sqlalchemy==1.4.23
twilio==6.64.0
scikit-learn==0.24.2
pycryptodome==3.11.0